+++ main/celerybeat-schedule.bak
@@ -1,0 +1,4 @@
+'entries', (0, 286)
+'__version__', (512, 15)
+'tz', (1024, 23)
+'utc_enabled', (1536, 4)
