+++ main/main/celery.py
@@ -1,7 +1,7 @@
 import os
 
 from celery import Celery
-
+from celery.schedules import crontab
 # set the default Django settings module for the 'celery' program.
 os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'main.settings')
 
@@ -16,7 +16,10 @@ app.config_from_object('django.conf:settings', namespace='CELERY')
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks()
 
-
-@app.task(bind=True)
-def debug_task(self):
-    print(f'Request: {self.request!r}')
+# Schedule to send new goods per week
+app.conf.beat_schedule = {
+    'weekly-news': {
+        'task': 'main.tasks.week_news_notifications',
+        'schedule': crontab(day_of_month='1,7,14,21'),
+    }
+}
