+++ requirements.txt
@@ -1,0 +1,81 @@
+alabaster==0.7.12
+amqp==2.6.1
+APScheduler==3.7.0
+asgiref==3.3.1
+atomicwrites==1.4.0
+attrs==20.3.0
+Babel==2.9.0
+billiard==3.6.4.0
+celery==4.4.7
+certifi==2020.12.5
+cffi==1.14.5
+chardet==4.0.0
+click==7.1.2
+click-didyoumean==0.0.3
+click-plugins==1.1.1
+click-repl==0.1.6
+colorama==0.4.4
+cryptography==3.4.7
+defusedxml==0.7.1
+Django==3.1.7
+django-allauth==0.44.0
+django-celery-beat==2.2.0
+django-celery-results==2.0.1
+django-ckeditor==6.0.0
+django-js-asset==1.2.2
+django-redis==4.12.1
+django-timezone-field==4.1.2
+dnspython==1.16.0
+docutils==0.16
+eventlet==0.30.2
+flake8==3.9.0
+flake8-import-order==0.18.1
+flower==0.9.7
+greenlet==1.0.0
+humanize==3.4.1
+idna==2.10
+imagesize==1.2.0
+iniconfig==1.1.1
+install==1.3.4
+Jinja2==2.11.3
+kombu==4.6.11
+MarkupSafe==1.1.1
+mccabe==0.6.1
+oauthlib==3.1.0
+packaging==20.9
+Pillow==8.1.2
+pluggy==0.13.1
+prometheus-client==0.8.0
+prompt-toolkit==3.0.18
+py==1.10.0
+pycodestyle==2.7.0
+pycparser==2.20
+pyflakes==2.3.0
+Pygments==2.8.1
+PyJWT==2.0.1
+pyparsing==2.4.7
+pytest==6.2.2
+python-crontab==2.5.1
+python-dateutil==2.8.1
+python3-openid==3.2.0
+pytz==2021.1
+redis==3.5.3
+requests==2.25.1
+requests-oauthlib==1.3.0
+six==1.15.0
+snowballstemmer==2.1.0
+sorl-thumbnail==12.7.0
+Sphinx==3.5.3
+sphinxcontrib-applehelp==1.0.2
+sphinxcontrib-devhelp==1.0.2
+sphinxcontrib-htmlhelp==1.0.3
+sphinxcontrib-jsmath==1.0.1
+sphinxcontrib-qthelp==1.0.3
+sphinxcontrib-serializinghtml==1.1.4
+sqlparse==0.4.1
+toml==0.10.2
+tornado==6.1
+tzlocal==2.1
+urllib3==1.26.4
+vine==1.3.0
+wcwidth==0.2.5
